Refactored components to own directory
