--- .eslintrc.json
@@ -1,10 +1,0 @@
-{
-    "extends": ["eslint:recommended"],
-    "parserOptions": {
-        "ecmaVersion": 6,
-        "sourceType": "module",
-        "ecmaFeatures": {
-            "jsx": true
-        }
-    }
-}
