--- package.json
@@ -1,25 +1,0 @@
-{
-  "name": "pokedex-front",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "bootstrap": "^4.1.3",
-    "react": "^16.5.2",
-    "react-dom": "^16.5.2",
-    "react-scripts": "1.1.5",
-    "reactstrap": "^6.4.0"
-  },
-  "devDependencies": {
-    "chai": "*",
-    "eslint": "^5.6.0",
-    "eslint-config-google": "^0.10.0",
-    "mocha": "*",
-    "sinon": "*"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject"
-  }
-}
